[consensus] tests
