+++ consensus/src/lib.rs
@@ -44,7 +44,7 @@ mod state_computer;
 mod state_computer_tests;
 mod state_replication;
 #[cfg(any(test, feature = "fuzzing"))]
-mod test_utils;
+pub(crate) mod test_utils;
 #[cfg(test)]
 mod twins;
 mod txn_notifier;
