+++ consensus/src/round_manager_tests/consensus_tests.rs
@@ -1,38 +1,20 @@
 // Copyright © Aptos Foundation
 // Parts of the project are originally copyright © Meta Platforms, Inc.
 // SPDX-License-Identifier: Apache-2.0
-
 use crate::{
-    block_storage::{pending_blocks::PendingBlocks, BlockReader, BlockStore},
+    block_storage::BlockReader,
     counters,
-    liveness::{
-        proposal_generator::{
-            ChainHealthBackoffConfig, PipelineBackpressureConfig, ProposalGenerator,
-        },
-        proposer_election::ProposerElection,
-        rotating_proposer_election::RotatingProposer,
-        round_state::{ExponentialTimeInterval, RoundState},
-    },
     metrics_safety_rules::MetricsSafetyRules,
-    network::{IncomingBlockRetrievalRequest, NetworkSender},
-    network_interface::{CommitMessage, ConsensusMsg, ConsensusNetworkClient, DIRECT_SEND, RPC},
+    network::IncomingBlockRetrievalRequest,
+    network_interface::ConsensusMsg,
     network_tests::{NetworkPlayground, TwinId},
-    payload_manager::DirectMempoolPayloadManager,
-    persistent_liveness_storage::RecoveryData,
-    pipeline::buffer_manager::OrderedBlocks,
-    round_manager::RoundManager,
-    test_utils::{
-        consensus_runtime, create_vec_signed_transactions,
-        mock_execution_client::MockExecutionClient, timed_block_on, MockOptQSPayloadProvider,
-        MockPastProposalStatusTracker, MockPayloadManager, MockStorage, TreeInserter,
+    round_manager::round_manager_tests::{
+        config_with_round_timeout_msg_disabled, start_replying_to_block_retreival, NodeSetup,
+        ProposalMsgType,
     },
-    util::time_service::{ClockTimeService, TimeService},
+    test_utils::{consensus_runtime, timed_block_on, TreeInserter},
 };
-use aptos_channels::{self, aptos_channel, message_queues::QueueStyle};
-use aptos_config::{
-    config::ConsensusConfig,
-    network_id::{NetworkId, PeerNetworkId},
-};
+use aptos_config::config::ConsensusConfig;
 use aptos_consensus_types::{
     block::{
         block_test_utils::{certificate_for_genesis, gen_test_certificate},
@@ -40,631 +22,22 @@ use aptos_consensus_types::{
     },
     block_retrieval::{BlockRetrievalRequest, BlockRetrievalRequestV1, BlockRetrievalStatus},
     common::{Author, Payload, Round},
-    order_vote_msg::OrderVoteMsg,
-    pipeline::commit_decision::CommitDecision,
+    opt_block_data,
     proposal_msg::ProposalMsg,
-    round_timeout::RoundTimeoutMsg,
     sync_info::SyncInfo,
     timeout_2chain::{TwoChainTimeout, TwoChainTimeoutWithPartialSignatures},
-    utils::PayloadTxnsSize,
-    vote_msg::VoteMsg,
 };
 use aptos_crypto::HashValue;
 use aptos_infallible::Mutex;
 use aptos_logger::prelude::info;
-use aptos_network::{
-    application::interface::NetworkClient,
-    peer_manager::{ConnectionRequestSender, PeerManagerRequestSender},
-    protocols::{
-        network,
-        network::{Event, NetworkEvents, NewNetworkEvents, NewNetworkSender},
-        wire::handshake::v1::ProtocolIdSet,
-    },
-    transport::ConnectionMetadata,
-    ProtocolId,
-};
+use aptos_network::{protocols::network::Event, ProtocolId};
 use aptos_safety_rules::{PersistentSafetyStorage, SafetyRulesManager};
 use aptos_secure_storage::Storage;
-use aptos_types::{
-    dkg::{real_dkg::RealDKG, DKGSessionMetadata, DKGTrait, DKGTranscript},
-    epoch_state::EpochState,
-    jwks::QuorumCertifiedUpdate,
-    ledger_info::LedgerInfo,
-    on_chain_config::{
-        ConsensusAlgorithmConfig, ConsensusConfigV1, OnChainConsensusConfig,
-        OnChainJWKConsensusConfig, OnChainRandomnessConfig, RandomnessConfigMoveStruct,
-        ValidatorTxnConfig, DEFAULT_WINDOW_SIZE,
-    },
-    transaction::SignedTransaction,
-    validator_signer::ValidatorSigner,
-    validator_txn::ValidatorTransaction,
-    validator_verifier::{
-        generate_validator_verifier, random_validator_verifier,
-        random_validator_verifier_with_voting_power, ValidatorConsensusInfoMoveStruct,
-        ValidatorVerifier,
-    },
-    waypoint::Waypoint,
-};
-use futures::{
-    channel::{mpsc, oneshot},
-    executor::block_on,
-    stream::select,
-    FutureExt, Stream, StreamExt,
-};
-use maplit::hashmap;
-use rand::{rngs::ThreadRng, thread_rng};
-use std::{
-    collections::VecDeque,
-    iter::FromIterator,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
-    time::Duration,
-};
-use tokio::{
-    runtime::{Handle, Runtime},
-    task::JoinHandle,
-    time::timeout,
-};
-
-/// Auxiliary struct that is setting up node environment for the test.
-pub struct NodeSetup {
-    block_store: Arc<BlockStore>,
-    round_manager: RoundManager,
-    storage: Arc<MockStorage>,
-    signer: ValidatorSigner,
-    proposers: Vec<Author>,
-    safety_rules_manager: SafetyRulesManager,
-    pending_network_events: Vec<Event<ConsensusMsg>>,
-    all_network_events: Box<dyn Stream<Item = Event<ConsensusMsg>> + Send + Unpin>,
-    ordered_blocks_events: mpsc::UnboundedReceiver<OrderedBlocks>,
-    mock_execution_client: Arc<MockExecutionClient>,
-    _state_sync_receiver: mpsc::UnboundedReceiver<Vec<SignedTransaction>>,
-    id: usize,
-    onchain_consensus_config: OnChainConsensusConfig,
-    local_consensus_config: ConsensusConfig,
-    onchain_randomness_config: OnChainRandomnessConfig,
-    onchain_jwk_consensus_config: OnChainJWKConsensusConfig,
-    vote_queue: VecDeque<VoteMsg>,
-    order_vote_queue: VecDeque<OrderVoteMsg>,
-    proposal_queue: VecDeque<ProposalMsg>,
-    round_timeout_queue: VecDeque<RoundTimeoutMsg>,
-    commit_decision_queue: VecDeque<CommitDecision>,
-}
-
-impl NodeSetup {
-    fn create_round_state(time_service: Arc<dyn TimeService>) -> RoundState {
-        let base_timeout = Duration::new(60, 0);
-        let time_interval = Box::new(ExponentialTimeInterval::fixed(base_timeout));
-        let (round_timeout_sender, _) = aptos_channels::new_test(1_024);
-        RoundState::new(time_interval, time_service, round_timeout_sender)
-    }
-
-    fn create_proposer_election(proposers: Vec<Author>) -> Arc<dyn ProposerElection + Send + Sync> {
-        Arc::new(RotatingProposer::new(proposers, 1))
-    }
-
-    fn create_nodes(
-        playground: &mut NetworkPlayground,
-        executor: Handle,
-        num_nodes: usize,
-        proposer_indices: Option<Vec<usize>>,
-        onchain_consensus_config: Option<OnChainConsensusConfig>,
-        local_consensus_config: Option<ConsensusConfig>,
-        onchain_randomness_config: Option<OnChainRandomnessConfig>,
-        onchain_jwk_consensus_config: Option<OnChainJWKConsensusConfig>,
-    ) -> Vec<Self> {
-        Self::create_nodes_with_validator_set(
-            playground,
-            executor,
-            num_nodes,
-            proposer_indices,
-            onchain_consensus_config,
-            local_consensus_config,
-            onchain_randomness_config,
-            onchain_jwk_consensus_config,
-            None,
-        )
-    }
-
-    fn create_nodes_with_validator_set(
-        playground: &mut NetworkPlayground,
-        executor: Handle,
-        num_nodes: usize,
-        proposer_indices: Option<Vec<usize>>,
-        onchain_consensus_config: Option<OnChainConsensusConfig>,
-        local_consensus_config: Option<ConsensusConfig>,
-        onchain_randomness_config: Option<OnChainRandomnessConfig>,
-        onchain_jwk_consensus_config: Option<OnChainJWKConsensusConfig>,
-        validator_set: Option<(Vec<ValidatorSigner>, ValidatorVerifier)>,
-    ) -> Vec<Self> {
-        let mut onchain_consensus_config = onchain_consensus_config.unwrap_or_default();
-        // With order votes feature, the validators additionally send order votes.
-        // next_proposal and next_vote functions could potentially break because of it.
-        if let OnChainConsensusConfig::V4 {
-            alg:
-                ConsensusAlgorithmConfig::JolteonV2 {
-                    main: _,
-                    quorum_store_enabled: _,
-                    order_vote_enabled,
-                },
-            vtxn: _,
-            window_size: _,
-        } = &mut onchain_consensus_config
-        {
-            *order_vote_enabled = false;
-        }
-        let onchain_randomness_config =
-            onchain_randomness_config.unwrap_or_else(OnChainRandomnessConfig::default_if_missing);
-        let onchain_jwk_consensus_config = onchain_jwk_consensus_config
-            .unwrap_or_else(OnChainJWKConsensusConfig::default_if_missing);
-        let local_consensus_config = local_consensus_config.unwrap_or_default();
-        let (signers, validators) =
-            validator_set.unwrap_or_else(|| random_validator_verifier(num_nodes, None, false));
-        let proposers = proposer_indices
-            .unwrap_or_else(|| vec![0])
-            .iter()
-            .map(|i| signers[*i].author())
-            .collect::<Vec<_>>();
-        let validator_set = (&validators).into();
-        let waypoint =
-            Waypoint::new_epoch_boundary(&LedgerInfo::mock_genesis(Some(validator_set))).unwrap();
-
-        let mut nodes = vec![];
-        // pre-initialize the mapping to avoid race conditions (peer try to broadcast to someone not added yet)
-        let peers_and_metadata = playground.peer_protocols();
-        for signer in signers.iter().take(num_nodes) {
-            let peer_id = signer.author();
-            let mut conn_meta = ConnectionMetadata::mock(peer_id);
-            conn_meta.application_protocols = ProtocolIdSet::from_iter([
-                ProtocolId::ConsensusDirectSendJson,
-                ProtocolId::ConsensusDirectSendBcs,
-                ProtocolId::ConsensusRpcBcs,
-            ]);
-            let peer_network_id = PeerNetworkId::new(NetworkId::Validator, peer_id);
-            peers_and_metadata
-                .insert_connection_metadata(peer_network_id, conn_meta)
-                .unwrap();
-        }
-        for (id, signer) in signers.iter().take(num_nodes).enumerate() {
-            let (initial_data, storage) = MockStorage::start_for_testing((&validators).into());
-
-            let safety_storage = PersistentSafetyStorage::initialize(
-                Storage::from(aptos_secure_storage::InMemoryStorage::new()),
-                signer.author(),
-                signer.private_key().clone(),
-                waypoint,
-                true,
-            );
-            let safety_rules_manager = SafetyRulesManager::new_local(safety_storage);
-
-            nodes.push(Self::new(
-                playground,
-                executor.clone(),
-                signer.to_owned(),
-                proposers.clone(),
-                storage,
-                initial_data,
-                safety_rules_manager,
-                id,
-                onchain_consensus_config.clone(),
-                local_consensus_config.clone(),
-                onchain_randomness_config.clone(),
-                onchain_jwk_consensus_config.clone(),
-            ));
-        }
-        nodes
-    }
-
-    fn new(
-        playground: &mut NetworkPlayground,
-        executor: Handle,
-        signer: ValidatorSigner,
-        proposers: Vec<Author>,
-        storage: Arc<MockStorage>,
-        initial_data: RecoveryData,
-        safety_rules_manager: SafetyRulesManager,
-        id: usize,
-        onchain_consensus_config: OnChainConsensusConfig,
-        local_consensus_config: ConsensusConfig,
-        onchain_randomness_config: OnChainRandomnessConfig,
-        onchain_jwk_consensus_config: OnChainJWKConsensusConfig,
-    ) -> Self {
-        let _entered_runtime = executor.enter();
-        let epoch_state = Arc::new(EpochState::new(1, storage.get_validator_set().into()));
-        let validators = epoch_state.verifier.clone();
-        let (network_reqs_tx, network_reqs_rx) = aptos_channel::new(QueueStyle::FIFO, 8, None);
-        let (connection_reqs_tx, _) = aptos_channel::new(QueueStyle::FIFO, 8, None);
-        let (consensus_tx, consensus_rx) = aptos_channel::new(QueueStyle::FIFO, 8, None);
-        let (_conn_mgr_reqs_tx, conn_mgr_reqs_rx) = aptos_channels::new_test(8);
-        let network_sender = network::NetworkSender::new(
-            PeerManagerRequestSender::new(network_reqs_tx),
-            ConnectionRequestSender::new(connection_reqs_tx),
-        );
-        let network_client = NetworkClient::new(
-            DIRECT_SEND.into(),
-            RPC.into(),
-            hashmap! {NetworkId::Validator => network_sender},
-            playground.peer_protocols(),
-        );
-        let consensus_network_client = ConsensusNetworkClient::new(network_client);
-        let network_events = NetworkEvents::new(consensus_rx, None, true);
-        let author = signer.author();
-
-        let twin_id = TwinId { id, author };
-
-        playground.add_node(twin_id, consensus_tx, network_reqs_rx, conn_mgr_reqs_rx);
-
-        let (self_sender, self_receiver) = aptos_channels::new_unbounded_test();
-        let network = Arc::new(NetworkSender::new(
-            author,
-            consensus_network_client,
-            self_sender,
-            validators,
-        ));
-
-        let all_network_events = Box::new(select(network_events, self_receiver));
-
-        let last_vote_sent = initial_data.last_vote();
-        let (ordered_blocks_tx, ordered_blocks_events) = mpsc::unbounded::<OrderedBlocks>();
-        let (state_sync_client, _state_sync_receiver) = mpsc::unbounded();
-        let mock_execution_client = Arc::new(MockExecutionClient::new(
-            state_sync_client.clone(),
-            ordered_blocks_tx.clone(),
-            Arc::clone(&storage),
-        ));
-        let time_service = Arc::new(ClockTimeService::new(executor));
-
-        let window_size = onchain_consensus_config.window_size();
-        let block_store = Arc::new(BlockStore::new(
-            storage.clone(),
-            initial_data,
-            mock_execution_client.clone(),
-            10, // max pruned blocks in mem
-            time_service.clone(),
-            10,
-            Arc::from(DirectMempoolPayloadManager::new()),
-            false,
-            window_size,
-            Arc::new(Mutex::new(PendingBlocks::new())),
-            None,
-        ));
-
-        let proposer_election = Self::create_proposer_election(proposers.clone());
-        let proposal_generator = ProposalGenerator::new(
-            author,
-            block_store.clone(),
-            Arc::new(MockPayloadManager::new(None)),
-            time_service.clone(),
-            Duration::ZERO,
-            PayloadTxnsSize::new(20, 1000),
-            10,
-            PayloadTxnsSize::new(5, 500),
-            10,
-            1,
-            Some(30_000),
-            PipelineBackpressureConfig::new_no_backoff(),
-            ChainHealthBackoffConfig::new_no_backoff(),
-            false,
-            onchain_consensus_config.effective_validator_txn_config(),
-            true,
-            Arc::new(MockOptQSPayloadProvider {}),
-        );
-
-        let round_state = Self::create_round_state(time_service);
-        let mut safety_rules =
-            MetricsSafetyRules::new(safety_rules_manager.client(), storage.clone());
-        safety_rules.perform_initialize().unwrap();
-
-        let (round_manager_tx, _) = aptos_channel::new(QueueStyle::LIFO, 1, None);
-
-        let (opt_proposal_loopback_tx, _) = aptos_channels::new_unbounded(
-            &counters::OP_COUNTERS.gauge("opt_proposal_loopback_queue"),
-        );
-
-        let local_config = local_consensus_config.clone();
-
-        let mut round_manager = RoundManager::new(
-            epoch_state,
-            Arc::clone(&block_store),
-            round_state,
-            proposer_election,
-            proposal_generator,
-            Arc::new(Mutex::new(safety_rules)),
-            network,
-            storage.clone(),
-            onchain_consensus_config.clone(),
-            round_manager_tx,
-            local_config,
-            onchain_randomness_config.clone(),
-            onchain_jwk_consensus_config.clone(),
-            None,
-            Arc::new(MockPastProposalStatusTracker {}),
-            opt_proposal_loopback_tx,
-        );
-        block_on(round_manager.init(last_vote_sent));
-        Self {
-            block_store,
-            round_manager,
-            storage,
-            signer,
-            proposers,
-            safety_rules_manager,
-            pending_network_events: Vec::new(),
-            all_network_events,
-            ordered_blocks_events,
-            mock_execution_client,
-            _state_sync_receiver,
-            id,
-            onchain_consensus_config,
-            local_consensus_config,
-            onchain_randomness_config,
-            onchain_jwk_consensus_config,
-            vote_queue: VecDeque::new(),
-            order_vote_queue: VecDeque::new(),
-            proposal_queue: VecDeque::new(),
-            round_timeout_queue: VecDeque::new(),
-            commit_decision_queue: VecDeque::new(),
-        }
-    }
-
-    pub fn restart(self, playground: &mut NetworkPlayground, executor: Handle) -> Self {
-        let recover_data = self
-            .storage
-            .try_start(
-                self.onchain_consensus_config.order_vote_enabled(),
-                self.onchain_consensus_config.window_size(),
-            )
-            .unwrap_or_else(|e| panic!("fail to restart due to: {}", e));
-        Self::new(
-            playground,
-            executor,
-            self.signer,
-            self.proposers,
-            self.storage,
-            recover_data,
-            self.safety_rules_manager,
-            self.id,
-            self.onchain_consensus_config.clone(),
-            self.local_consensus_config.clone(),
-            self.onchain_randomness_config.clone(),
-            self.onchain_jwk_consensus_config.clone(),
-        )
-    }
-
-    pub fn identity_desc(&self) -> String {
-        format!("{} [{}]", self.id, self.signer.author())
-    }
-
-    fn poll_next_network_event(&mut self) -> Option<Event<ConsensusMsg>> {
-        if !self.pending_network_events.is_empty() {
-            Some(self.pending_network_events.remove(0))
-        } else {
-            self.all_network_events
-                .next()
-                .now_or_never()
-                .map(|v| v.unwrap())
-        }
-    }
-
-    pub async fn next_network_event(&mut self) -> Event<ConsensusMsg> {
-        if !self.pending_network_events.is_empty() {
-            self.pending_network_events.remove(0)
-        } else {
-            self.all_network_events.next().await.unwrap()
-        }
-    }
-
-    pub async fn next_network_message(&mut self) {
-        let consensus_msg = match self.next_network_event().await {
-            Event::Message(_, msg) => msg,
-            Event::RpcRequest(_, msg, _, _) if matches!(msg, ConsensusMsg::CommitMessage(_)) => msg,
-            Event::RpcRequest(_, msg, _, _) => {
-                panic!(
-                    "Unexpected event, got RpcRequest, expected Message: {:?} on node {}",
-                    msg,
-                    self.identity_desc()
-                )
-            },
-        };
-
-        match consensus_msg {
-            ConsensusMsg::ProposalMsg(proposal) => {
-                self.proposal_queue.push_back(*proposal);
-            },
-            ConsensusMsg::VoteMsg(vote) => {
-                self.vote_queue.push_back(*vote);
-            },
-            ConsensusMsg::OrderVoteMsg(order_vote) => {
-                self.order_vote_queue.push_back(*order_vote);
-            },
-            ConsensusMsg::RoundTimeoutMsg(round_timeout) => {
-                self.round_timeout_queue.push_back(*round_timeout);
-            },
-            ConsensusMsg::CommitDecisionMsg(commit_decision) => {
-                self.commit_decision_queue.push_back(*commit_decision);
-            },
-            ConsensusMsg::CommitMessage(d) if matches!(*d, CommitMessage::Decision(_)) => {
-                match *d {
-                    CommitMessage::Decision(commit_decision) => {
-                        self.commit_decision_queue.push_back(commit_decision);
-                    },
-                    _ => unreachable!(),
-                }
-            },
-            msg => panic!(
-                "Unexpected Consensus Message: {:?} on node {}",
-                msg,
-                self.identity_desc()
-            ),
-        }
-    }
-
-    pub fn no_next_msg(&mut self) {
-        match self.poll_next_network_event() {
-            Some(Event::RpcRequest(_, msg, _, _)) | Some(Event::Message(_, msg)) => panic!(
-                "Unexpected Consensus Message: {:?} on node {}",
-                msg,
-                self.identity_desc()
-            ),
-            None => {},
-        }
-    }
-
-    pub async fn next_proposal(&mut self) -> ProposalMsg {
-        while self.proposal_queue.is_empty() {
-            self.next_network_message().await;
-        }
-        self.proposal_queue.pop_front().unwrap()
-    }
-
-    pub async fn next_vote(&mut self) -> VoteMsg {
-        while self.vote_queue.is_empty() {
-            self.next_network_message().await;
-        }
-        self.vote_queue.pop_front().unwrap()
-    }
-
-    #[allow(unused)]
-    pub async fn next_order_vote(&mut self) -> OrderVoteMsg {
-        while self.order_vote_queue.is_empty() {
-            self.next_network_message().await;
-        }
-        self.order_vote_queue.pop_front().unwrap()
-    }
-
-    pub async fn next_timeout(&mut self) -> RoundTimeoutMsg {
-        while self.round_timeout_queue.is_empty() {
-            self.next_network_message().await;
-        }
-        self.round_timeout_queue.pop_front().unwrap()
-    }
-
-    pub async fn next_commit_decision(&mut self) -> CommitDecision {
-        while self.commit_decision_queue.is_empty() {
-            self.next_network_message().await;
-        }
-        self.commit_decision_queue.pop_front().unwrap()
-    }
-
-    pub async fn poll_block_retrieval(&mut self) -> Option<IncomingBlockRetrievalRequest> {
-        match self.poll_next_network_event() {
-            Some(Event::RpcRequest(_, msg, protocol, response_sender)) => match msg {
-                ConsensusMsg::DeprecatedBlockRetrievalRequest(v) => {
-                    Some(IncomingBlockRetrievalRequest {
-                        req: BlockRetrievalRequest::V1(*v),
-                        protocol,
-                        response_sender,
-                    })
-                },
-                ConsensusMsg::BlockRetrievalRequest(v) => Some(IncomingBlockRetrievalRequest {
-                    req: *v,
-                    protocol,
-                    response_sender,
-                }),
-                msg => panic!(
-                    "Unexpected Consensus Message: {:?} on node {}",
-                    msg,
-                    self.identity_desc()
-                ),
-            },
-            Some(Event::Message(_, msg)) => panic!(
-                "Unexpected Consensus Message: {:?} on node {}",
-                msg,
-                self.identity_desc()
-            ),
-            None => None,
-        }
-    }
-
-    pub fn no_next_ordered(&mut self) {
-        if self.ordered_blocks_events.next().now_or_never().is_some() {
-            panic!("Unexpected Ordered Blocks Event");
-        }
-    }
-
-    pub async fn commit_next_ordered(&mut self, expected_rounds: &[Round]) {
-        info!(
-            "Starting commit_next_ordered to wait for {:?} on node {:?}",
-            expected_rounds,
-            self.identity_desc()
-        );
-        let ordered_blocks = self.ordered_blocks_events.next().await.unwrap();
-        let rounds = ordered_blocks
-            .ordered_blocks
-            .iter()
-            .map(|b| b.round())
-            .collect::<Vec<_>>();
-        assert_eq!(&rounds, expected_rounds);
-        self.mock_execution_client
-            .commit_to_storage(ordered_blocks)
-            .await
-            .unwrap();
-    }
-}
-
-fn config_with_round_timeout_msg_disabled() -> ConsensusConfig {
-    // Disable RoundTimeoutMsg to unless expliclity enabled.
-    ConsensusConfig {
-        enable_round_timeout_msg: false,
-        ..Default::default()
-    }
-}
-
-fn start_replying_to_block_retreival(nodes: Vec<NodeSetup>) -> ReplyingRPCHandle {
-    let done = Arc::new(AtomicBool::new(false));
-    let mut handles = Vec::new();
-    for mut node in nodes.into_iter() {
-        let done_clone = done.clone();
-        handles.push(tokio::spawn(async move {
-            while !done_clone.load(Ordering::Relaxed) {
-                info!("Asking for RPC request on {:?}", node.identity_desc());
-                let maybe_request = node.poll_block_retrieval().await;
-                if let Some(request) = maybe_request {
-                    info!(
-                        "RPC request received: {:?} on {:?}",
-                        request,
-                        node.identity_desc()
-                    );
-                    let wrapped_request = IncomingBlockRetrievalRequest {
-                        req: request.req,
-                        protocol: request.protocol,
-                        response_sender: request.response_sender,
-                    };
-                    node.block_store
-                        .process_block_retrieval(wrapped_request)
-                        .await
-                        .unwrap();
-                } else {
-                    tokio::time::sleep(Duration::from_millis(50)).await;
-                }
-            }
-            node
-        }));
-    }
-    ReplyingRPCHandle { handles, done }
-}
-
-struct ReplyingRPCHandle {
-    handles: Vec<JoinHandle<NodeSetup>>,
-    done: Arc<AtomicBool>,
-}
-
-impl ReplyingRPCHandle {
-    async fn join(self) -> Vec<NodeSetup> {
-        self.done.store(true, Ordering::Relaxed);
-        let mut result = Vec::new();
-        for handle in self.handles.into_iter() {
-            result.push(handle.await.unwrap());
-        }
-        info!(
-            "joined nodes in order: {:?}",
-            result.iter().map(|v| v.id).collect::<Vec<_>>()
-        );
-        result
-    }
-}
+use aptos_types::validator_verifier::generate_validator_verifier;
+use futures::channel::oneshot;
+use std::{sync::Arc, time::Duration};
+use tokio::{runtime::Runtime, time::timeout};
+use tokio_stream::StreamExt;
 
 fn process_and_vote_on_proposal(
     runtime: &Runtime,
@@ -688,29 +61,72 @@ fn process_and_vote_on_proposal(
         info!("Waiting on next_proposal on node {}", node.identity_desc());
         if down_nodes.contains(&node.id) {
             // Drop the proposal on down nodes
-            timed_block_on(runtime, node.next_proposal());
+            timed_block_on(runtime, node.next_opt_or_normal_proposal());
             info!("Dropping proposal on down node {}", node.identity_desc());
         } else {
             // Proccess proposal on other nodes
-            let proposal_msg = timed_block_on(runtime, node.next_proposal());
-            info!("Processing proposal on {}", node.identity_desc());
+            let proposal_msg_type = timed_block_on(runtime, node.next_opt_or_normal_proposal());
+            match proposal_msg_type {
+                ProposalMsgType::Normal(proposal_msg) => {
+                    info!("Processing proposal on {}", node.identity_desc());
 
-            assert_eq!(proposal_msg.proposal().round(), expected_round);
-            assert_eq!(
-                proposal_msg.sync_info().highest_ordered_round(),
-                expected_qc_ordered_round
-            );
-            assert_eq!(
-                proposal_msg.sync_info().highest_commit_round(),
-                expected_qc_committed_round
-            );
+                    assert_eq!(proposal_msg.proposal().round(), expected_round);
+                    assert_eq!(
+                        proposal_msg.sync_info().highest_ordered_round(),
+                        expected_qc_ordered_round
+                    );
+                    assert_eq!(
+                        proposal_msg.sync_info().highest_commit_round(),
+                        expected_qc_committed_round
+                    );
 
-            timed_block_on(
-                runtime,
-                node.round_manager.process_proposal_msg(proposal_msg),
-            )
-            .unwrap();
-            info!("Finish process proposal on {}", node.identity_desc());
+                    timed_block_on(
+                        runtime,
+                        node.round_manager.process_proposal_msg(proposal_msg),
+                    )
+                    .unwrap();
+                    info!("Finish process proposal on {}", node.identity_desc());
+                },
+                ProposalMsgType::Optimistic(opt_proposal_msg) => {
+                    info!("Processing opt proposal on {}", node.identity_desc());
+
+                    assert_eq!(opt_proposal_msg.round(), expected_round);
+                    assert_eq!(
+                        opt_proposal_msg.sync_info().highest_ordered_round(),
+                        expected_qc_ordered_round.saturating_sub(1)
+                    );
+                    assert_eq!(
+                        opt_proposal_msg.sync_info().highest_commit_round(),
+                        expected_qc_committed_round
+                    );
+
+                    timed_block_on(
+                        runtime,
+                        node.round_manager
+                            .process_opt_proposal_msg(opt_proposal_msg),
+                    )
+                    .unwrap();
+                    info!("Finish process opt proposal on {}", node.identity_desc());
+
+                    info!(
+                        "Processing proposal (from opt proposal) on {}",
+                        node.identity_desc()
+                    );
+
+                    let opt_block_data =
+                        timed_block_on(runtime, node.processed_opt_proposal_rx.next()).unwrap();
+                    timed_block_on(
+                        runtime,
+                        node.round_manager.process_opt_proposal(opt_block_data),
+                    )
+                    .unwrap();
+
+                    info!(
+                        "Finish process proposal (from opt proposal) on {}",
+                        node.identity_desc()
+                    );
+                },
+            }
             num_votes += 1;
         }
     }
@@ -785,16 +201,30 @@ fn new_round_on_quorum_cert() {
             .process_proposal_msg(proposal_msg)
             .await
             .unwrap();
+
         let vote_msg = node.next_vote().await;
         // Adding vote to form a QC
         node.round_manager.process_vote_msg(vote_msg).await.unwrap();
 
         // round 2 should start
-        let proposal_msg = node.next_proposal().await;
-        let proposal = proposal_msg.proposal();
-        assert_eq!(proposal.round(), 2);
-        assert_eq!(proposal.parent_id(), b1_id);
-        assert_eq!(proposal.quorum_cert().certified_block().id(), b1_id);
+        let proposal_msg_type = node.next_opt_or_normal_proposal().await;
+        match proposal_msg_type {
+            ProposalMsgType::Normal(proposal_msg) => {
+                let proposal = proposal_msg.proposal();
+                assert_eq!(proposal.round(), 2);
+                assert_eq!(proposal.parent_id(), b1_id);
+                assert_eq!(proposal.quorum_cert().certified_block().id(), b1_id);
+            },
+            ProposalMsgType::Optimistic(opt_proposal_msg) => {
+                let proposal = opt_proposal_msg.block_data();
+                assert_eq!(proposal.round(), 2);
+                assert_eq!(proposal.parent_id(), b1_id);
+                assert_eq!(
+                    proposal.grandparent_qc().certified_block().id(),
+                    genesis.id()
+                );
+            },
+        }
     });
 }
 
@@ -869,9 +299,7 @@ fn delay_proposal_processing_in_sync_only() {
         node.next_proposal().await;
 
         // Set sync only to true so that new proposal processing is delayed.
-        node.round_manager
-            .block_store
-            .set_back_pressure_for_test(true);
+        node.block_store.set_back_pressure_for_test(true);
         let proposal = Block::new_proposal(
             Payload::empty(false, true),
             1,
@@ -893,9 +321,7 @@ fn delay_proposal_processing_in_sync_only() {
             .unwrap_err();
 
         // Clear the sync only mode and process verified proposal and ensure it is processed now
-        node.round_manager
-            .block_store
-            .set_back_pressure_for_test(false);
+        node.block_store.set_back_pressure_for_test(false);
 
         node.round_manager
             .process_verified_proposal(proposal)
@@ -1813,14 +1239,30 @@ fn safety_rules_crash() {
 
     timed_block_on(&runtime, async {
         for _ in 0..2 {
-            let proposal_msg = node.next_proposal().await;
+            let proposal_msg_type = node.next_opt_or_normal_proposal().await;
 
             reset_safety_rules(&mut node);
-            // construct_and_sign_vote
-            node.round_manager
-                .process_proposal_msg(proposal_msg)
-                .await
-                .unwrap();
+
+            match proposal_msg_type {
+                ProposalMsgType::Normal(proposal_msg) => {
+                    // construct_and_sign_vote
+                    node.round_manager
+                        .process_proposal_msg(proposal_msg)
+                        .await
+                        .unwrap();
+                },
+                ProposalMsgType::Optimistic(opt_proposal_msg) => {
+                    node.round_manager
+                        .process_opt_proposal_msg(opt_proposal_msg)
+                        .await
+                        .unwrap();
+                    let opt_block_data = node.processed_opt_proposal_rx.next().await.unwrap();
+                    node.round_manager
+                        .process_opt_proposal(opt_block_data)
+                        .await
+                        .unwrap();
+                },
+            }
 
             let vote_msg = node.next_vote().await;
 
@@ -1839,7 +1281,7 @@ fn safety_rules_crash() {
         }
 
         // verify the last sign proposal happened
-        node.next_proposal().await;
+        node.next_opt_or_normal_proposal().await;
     });
 }
 
@@ -2073,7 +1515,7 @@ fn block_retrieval_test() {
 
         // Drain the queue on other nodes
         for node in nodes.iter_mut() {
-            let _ = node.next_proposal().await;
+            let _ = node.next_opt_or_normal_proposal().await;
         }
 
         info!(
@@ -2081,12 +1523,23 @@ fn block_retrieval_test() {
             behind_node.identity_desc()
         );
         let handle = start_replying_to_block_retreival(nodes);
-        let proposal_msg = behind_node.next_proposal().await;
-        behind_node
-            .round_manager
-            .process_proposal_msg(proposal_msg)
-            .await
-            .unwrap();
+
+        let proposal_msg_type = behind_node.next_opt_or_normal_proposal().await;
+        info!("got proposal msg: {:?}", proposal_msg_type);
+        match proposal_msg_type {
+            ProposalMsgType::Normal(proposal_msg) => behind_node
+                .round_manager
+                .process_proposal_msg(proposal_msg)
+                .await
+                .unwrap(),
+            ProposalMsgType::Optimistic(opt_proposal_msg) => {
+                behind_node
+                    .round_manager
+                    .process_opt_proposal_msg(opt_proposal_msg)
+                    .await
+                    .unwrap();
+            },
+        }
 
         handle.join().await;
     });
@@ -2143,7 +1596,7 @@ fn block_retrieval_timeout_test() {
 
         // Drain the queue on other nodes
         for node in nodes.iter_mut() {
-            let _ = node.next_proposal().await;
+            let _ = node.next_opt_or_normal_proposal().await;
         }
 
         info!(
@@ -2151,12 +1604,23 @@ fn block_retrieval_timeout_test() {
             behind_node.identity_desc()
         );
 
-        let proposal_msg = behind_node.next_proposal().await;
-        behind_node
-            .round_manager
-            .process_proposal_msg(proposal_msg)
-            .await
-            .unwrap_err();
+        let proposal_msg_type = behind_node.next_opt_or_normal_proposal().await;
+        match proposal_msg_type {
+            ProposalMsgType::Normal(proposal_msg) => {
+                behind_node
+                    .round_manager
+                    .process_proposal_msg(proposal_msg)
+                    .await
+                    .unwrap_err();
+            },
+            ProposalMsgType::Optimistic(opt_proposal_msg) => {
+                behind_node
+                    .round_manager
+                    .process_opt_proposal_msg(opt_proposal_msg)
+                    .await
+                    .unwrap_err();
+            },
+        }
     });
 }
 
@@ -2414,367 +1878,4 @@ pub fn forking_retrieval_test() {
         7,
         3,
     );
-}
-
-#[test]
-/// If ProposalExt feature is disabled, ProposalExt should be rejected
-/// No votes are sent, but the block is still added to the block tree.
-fn no_vote_on_proposal_ext_when_feature_disabled() {
-    let runtime = consensus_runtime();
-    let mut playground = NetworkPlayground::new(runtime.handle().clone());
-    // In order to observe the votes we're going to check proposal processing on the non-proposer
-    // node (which will send the votes to the proposer).
-    let mut nodes = NodeSetup::create_nodes(
-        &mut playground,
-        runtime.handle().clone(),
-        1,
-        None,
-        None,
-        None,
-        None,
-        None,
-    );
-    let node = &mut nodes[0];
-    let genesis_qc = certificate_for_genesis();
-
-    let invalid_block = Block::new_proposal_ext(
-        vec![ValidatorTransaction::dummy(vec![0xFF]); 5],
-        Payload::empty(false, true),
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    let valid_block = Block::new_proposal(
-        Payload::empty(false, true),
-        1,
-        1,
-        genesis_qc,
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    timed_block_on(&runtime, async {
-        // clear the message queue
-        node.next_proposal().await;
-
-        assert!(node
-            .round_manager
-            .process_proposal(invalid_block)
-            .await
-            .is_err());
-
-        assert!(node
-            .round_manager
-            .process_proposal(valid_block)
-            .await
-            .is_ok());
-    });
-}
-
-#[test]
-fn no_vote_on_proposal_with_unexpected_vtxns() {
-    let vtxns = vec![ValidatorTransaction::ObservedJWKUpdate(
-        QuorumCertifiedUpdate::dummy(),
-    )];
-
-    assert_process_proposal_result(
-        None,
-        None,
-        Some(OnChainJWKConsensusConfig::default_disabled()),
-        vtxns.clone(),
-        false,
-    );
-
-    assert_process_proposal_result(
-        None,
-        None,
-        Some(OnChainJWKConsensusConfig::default_enabled()),
-        vtxns,
-        true,
-    );
-}
-
-#[test]
-fn no_vote_on_proposal_with_uncertified_dkg_result() {
-    test_dkg_result_handling(
-        &[25_000_000; 4],
-        1,
-        RealDKG::sample_secret_and_generate_transcript,
-        false,
-    );
-}
-
-#[test]
-fn no_vote_on_proposal_with_inconsistent_secret_dkg_result() {
-    test_dkg_result_handling(
-        &[10_000_000, 70_000_000, 10_000_000, 10_000_000],
-        1,
-        RealDKG::generate_transcript_for_inconsistent_secrets,
-        false,
-    );
-}
-
-#[test]
-fn no_vote_on_proposal_with_dup_dealers_in_dkg_transcript() {
-    test_dkg_result_handling(
-        &[10_000_000, 40_000_000, 10_000_000, 40_000_000],
-        1,
-        RealDKG::deal_twice_and_aggregate,
-        false,
-    );
-}
-
-#[test]
-fn vote_on_proposal_with_valid_dkg_result() {
-    test_dkg_result_handling(
-        &[10_000_000, 70_000_000, 10_000_000, 10_000_000],
-        1,
-        RealDKG::sample_secret_and_generate_transcript,
-        true,
-    );
-}
-
-fn test_dkg_result_handling<F>(
-    voting_powers: &[u64],
-    dealer_idx: usize,
-    trx_gen_func: F,
-    should_accept: bool,
-) where
-    F: Fn(
-        &mut ThreadRng,
-        &<RealDKG as DKGTrait>::PublicParams,
-        u64,
-        &<RealDKG as DKGTrait>::DealerPrivateKey,
-    ) -> <RealDKG as DKGTrait>::Transcript,
-{
-    let mut rng = thread_rng();
-    let epoch = 123;
-    let num_validators = voting_powers.len();
-    let (signers, verifier) =
-        random_validator_verifier_with_voting_power(num_validators, None, false, voting_powers);
-    let validator_set: Vec<ValidatorConsensusInfoMoveStruct> = verifier
-        .validator_infos
-        .clone()
-        .into_iter()
-        .map(ValidatorConsensusInfoMoveStruct::from)
-        .collect();
-
-    let dkg_session_metadata = DKGSessionMetadata {
-        dealer_epoch: epoch,
-        randomness_config: RandomnessConfigMoveStruct::from(
-            OnChainRandomnessConfig::default_enabled(),
-        ),
-        dealer_validator_set: validator_set.clone(),
-        target_validator_set: validator_set,
-    };
-    let public_params = RealDKG::new_public_params(&dkg_session_metadata);
-    let trx = trx_gen_func(
-        &mut rng,
-        &public_params,
-        dealer_idx as u64,
-        signers[dealer_idx].private_key(),
-    );
-    let trx_bytes = bcs::to_bytes(&trx).unwrap();
-    let vtxns = vec![ValidatorTransaction::DKGResult(DKGTranscript::new(
-        epoch,
-        verifier.get_ordered_account_addresses()[dealer_idx],
-        trx_bytes,
-    ))];
-
-    assert_process_proposal_result(
-        Some((signers, verifier)),
-        Some(OnChainRandomnessConfig::default_enabled()),
-        Some(OnChainJWKConsensusConfig::default_enabled()),
-        vtxns.clone(),
-        should_accept,
-    );
-}
-
-/// Setup a node with default configs and an optional `Features` override.
-/// Create a block, fill it with the given vtxns, and process it with the `RoundManager` from the setup.
-/// Assert the processing result.
-fn assert_process_proposal_result(
-    validator_set: Option<(Vec<ValidatorSigner>, ValidatorVerifier)>,
-    randomness_config: Option<OnChainRandomnessConfig>,
-    jwk_consensus_config: Option<OnChainJWKConsensusConfig>,
-    vtxns: Vec<ValidatorTransaction>,
-    expected_result: bool,
-) {
-    let runtime = consensus_runtime();
-    let mut playground = NetworkPlayground::new(runtime.handle().clone());
-    let mut nodes = NodeSetup::create_nodes_with_validator_set(
-        &mut playground,
-        runtime.handle().clone(),
-        1,
-        None,
-        Some(OnChainConsensusConfig::default_for_genesis()),
-        None,
-        randomness_config,
-        jwk_consensus_config,
-        validator_set,
-    );
-
-    let node = &mut nodes[0];
-    let genesis_qc = certificate_for_genesis();
-    let block = Block::new_proposal_ext(
-        vtxns,
-        Payload::empty(false, true),
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    timed_block_on(&runtime, async {
-        // clear the message queue
-        node.next_proposal().await;
-
-        assert_eq!(
-            expected_result,
-            node.round_manager
-                .process_proposal(block.clone())
-                .await
-                .is_ok()
-        );
-    });
-}
-
-#[ignore]
-#[test]
-/// If receiving txn num/block size limit is exceeded, ProposalExt should be rejected.
-/// TODO: re-implement dummy vtxn and re-enable.
-fn no_vote_on_proposal_ext_when_receiving_limit_exceeded() {
-    let runtime = consensus_runtime();
-    let mut playground = NetworkPlayground::new(runtime.handle().clone());
-
-    let alg_config = ConsensusAlgorithmConfig::JolteonV2 {
-        main: ConsensusConfigV1::default(),
-        quorum_store_enabled: true,
-        order_vote_enabled: false,
-    };
-    let vtxn_config = ValidatorTxnConfig::V1 {
-        per_block_limit_txn_count: 5,
-        per_block_limit_total_bytes: 400,
-    };
-
-    let local_config = ConsensusConfig {
-        max_receiving_block_txns: 10,
-        max_receiving_block_bytes: 800,
-        ..Default::default()
-    };
-
-    let randomness_config = OnChainRandomnessConfig::default_enabled();
-    let mut nodes = NodeSetup::create_nodes(
-        &mut playground,
-        runtime.handle().clone(),
-        1,
-        None,
-        Some(OnChainConsensusConfig::V4 {
-            alg: alg_config,
-            vtxn: vtxn_config,
-            window_size: DEFAULT_WINDOW_SIZE,
-        }),
-        Some(local_config),
-        Some(randomness_config),
-        None,
-    );
-    let node = &mut nodes[0];
-    let genesis_qc = certificate_for_genesis();
-
-    let block_too_many_txns = Block::new_proposal_ext(
-        vec![],
-        Payload::DirectMempool(create_vec_signed_transactions(11)),
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    let block_too_many_vtxns = Block::new_proposal_ext(
-        vec![ValidatorTransaction::dummy(vec![0xFF; 20]); 6],
-        Payload::DirectMempool(create_vec_signed_transactions(4)),
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    let block_too_large = Block::new_proposal_ext(
-        vec![ValidatorTransaction::dummy(vec![0xFF; 200]); 1], // total_bytes >= 200 * 1 = 200
-        Payload::DirectMempool(create_vec_signed_transactions(9)), // = total_bytes >= 69 * 9 = 621
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    let block_vtxns_too_large = Block::new_proposal_ext(
-        vec![ValidatorTransaction::dummy(vec![0xFF; 200]); 5], // total_bytes >= 200 * 5 = 1000
-        Payload::empty(false, true),
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    let valid_block = Block::new_proposal_ext(
-        vec![ValidatorTransaction::dummy(vec![0xFF; 20]); 5], // total_bytes >= 60 * 5 = 300
-        Payload::DirectMempool(create_vec_signed_transactions(5)), // total_bytes >= 69 * 5 = 345
-        1,
-        1,
-        genesis_qc.clone(),
-        &node.signer,
-        Vec::new(),
-    )
-    .unwrap();
-
-    timed_block_on(&runtime, async {
-        // clear the message queue
-        node.next_proposal().await;
-
-        assert!(node
-            .round_manager
-            .process_proposal(block_too_many_txns)
-            .await
-            .is_err());
-
-        assert!(node
-            .round_manager
-            .process_proposal(block_too_many_vtxns)
-            .await
-            .is_err());
-
-        assert!(node
-            .round_manager
-            .process_proposal(block_too_large)
-            .await
-            .is_err());
-
-        assert!(node
-            .round_manager
-            .process_proposal(block_vtxns_too_large)
-            .await
-            .is_err());
-
-        assert!(node
-            .round_manager
-            .process_proposal(valid_block)
-            .await
-            .is_ok());
-    });
 }
